third_party/blink/web_tests/fast/forms/radionodelist-image-type-expected.txt

RadioNodeList should respond to type attribute change.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initially, there are three input elements with type=text.
nodeList = form.elements.namedItem('someName');
PASS nodeList.length is 3
PASS document.getElementsByTagName('input')[2].type = 'image'; nodeList.length is 2
PASS document.getElementsByTagName('input')[2].type = 'radio'; nodeList.length is 3
PASS successfullyParsed is true

TEST COMPLETE
